refactor: move decay to behavior
